+++ iocore/aio/test_AIO.cc
@@ -22,7 +22,7 @@
  */
 
 #include "P_AIO.h"
-#include "InkAPIInternal.h"
+#include "api/InkAPIInternal.h"
 #include "tscore/ink_hw.h"
 #include "tscore/I_Layout.h"
 #include "tscore/TSSystemState.h"
