+++ proxy/Transform.h
@@ -25,7 +25,7 @@
 
 #include "I_EventSystem.h"
 #include "HTTP.h"
-#include "InkAPIInternal.h"
+#include "api/InkAPIInternal.h"
 
 #define TRANSFORM_READ_READY (TRANSFORM_EVENTS_START + 0)
